Split large code file functions
