+++ packages/features/bookings/lib/handleSeats/reschedule/rescheduleSeatedBooking.ts
@@ -6,7 +6,7 @@ import prisma from "@calcom/prisma";
 import { BookingStatus } from "@calcom/prisma/enums";
 import type { Person } from "@calcom/types/Calendar";
 
-import { refreshCredentials } from "../../handleNewBooking";
+import { refreshCredentials } from "../../../../../../modules/credentials/refreshCredentials";
 import type { createLoggerWithEventDetails } from "../../handleNewBooking";
 import type {
   HandleSeatsResultBooking,
